+++ src/datasources/alarms/components/editors/alarms-count/AlarmsCountQueryEditor.test.tsx
@@ -3,7 +3,7 @@ import { act, render, screen } from '@testing-library/react';
 import { AlarmsCountQueryEditor } from './AlarmsCountQueryEditor';
 import { QueryType } from 'datasources/alarms/types/types';
 import { AlarmsCountQuery } from 'datasources/alarms/types/AlarmsCount.types';
-import { AlarmsCountDataSource } from 'datasources/alarms/query-type-handlers/alarms-count/AlarmsCountDataSource';
+import { AlarmsCountQueryHandler } from 'datasources/alarms/query-type-handlers/alarms-count/AlarmsCountQueryHandler';
 
 const mockHandleQueryChange = jest.fn();
 const mockGlobalVars = [{ label: '$var1', value: '$value1' }];
@@ -15,7 +15,7 @@ const mockDatasource = {
       ['2', { id: '2', name: 'AnotherWorkspaceName' }],
     ])
   ),
-} as unknown as AlarmsCountDataSource;
+} as unknown as AlarmsCountQueryHandler;
 
 const defaultProps = {
   query: {